requirements.txt
requirements according to os's

Linux:
    portaudio19-dev

Windows:
    pywin32

MacOs:
    AppKit

Common:
    pipwin
    pyaudio
    SpeechRecognition
    pynput
    python3-testresources
    google-api-python-client
    difflib